+++ src/cli/commands/identity.ts
@@ -1,5 +1,6 @@
 import { Command } from "commander";
-import { IdentityManagementClient, IdentityManagementModels } from "../../api/sdk";
+import { IdentityManagementClient, IdentityManagementModels, TenantManagementClient } from "../../api/sdk";
+import { throwError } from "../../api/utils";
 import {
     adjustColor,
     errorLog,
@@ -27,23 +28,26 @@ export default (program: Command) => {
         )
         .option("-u, --user [user]", "user name")
         .option("-g, --group [group]", "user group")
-        .option("-l, --role [role]", "user role")
+        .option("-r, --role [role]", "user role")
+        .option("-s, --subtenant <subtenant>", "subtenant for user")
         .option("-m, --membergroup [membergroup]", "member group")
+        .option("-l, --memberrole [memberrole]", "member role")
         .option("-t, --meta", "include meta information (ids, login details etc.)")
-        .option("-r, --raw", "don't automatically preceed group names with mdsp_usergroup")
+        .option("-w, --raw", "don't automatically preceed group names with mdsp_usergroup or role with mdsp_customrole")
         .option("-k, --passkey <passkey>", "passkey")
-        .option("-v, --verbose", "verbose output (includes assigned users or groups and roles)")
-        .description(color("manage mindsphere users and groups *"))
+        .option("-v, --verbose", "verbose output")
+        .description(color("manage mindsphere users, groups and roles *"))
         .action((options) => {
             (async () => {
                 try {
-                    // checkRequiredParameters(options);
+                    checkRequiredParameters(options);
                     const sdk = getSdk(options);
                     color = adjustColor(color, options);
                     homeDirLog(options.verbose, color);
                     proxyLog(options.verbose, color);
 
                     const iam = sdk.GetIdentityManagementClient();
+                    const tenantManagement = sdk.GetTenantManagementClient();
 
                     const users = async () => await getAllUsers(iam, options);
                     const groups = async () => await getAllGroups(iam, options);
@@ -63,16 +67,71 @@ export default (program: Command) => {
                         options.role &&
                         printRoles(await users(), await groups(), await roles(), options);
 
-                    // options.mode === "create" && options.user && (await createUser(iam, options));
-                    // options.mode === "create" && options.group && (await createGroup(iam, options));
+                    // * create commands
 
-                    // options.mode === "delete" && options.user && (await deleteUser(iam, options));
-                    // options.mode === "delete" && options.group && (await deleteGroup(iam, options));
+                    options.mode === "create" && options.user && (await createUser(iam, tenantManagement, options));
+                    options.mode === "create" && options.group && (await createGroup(iam, options));
+                    options.mode === "create" && options.role && (await createRole(iam, options));
 
-                    // options.mode === "assign" && options.role && (await assignRole(iam, options));
-                    // options.mode === "assign" && !options.role && (await assign(iam, options));
+                    // * delete commands
 
-                    // options.mode === "remove" && (await remove(iam, options));
+                    options.mode === "delete" && options.user && (await deleteUser(iam, options));
+                    options.mode === "delete" && options.group && (await deleteGroup(iam, options));
+                    options.mode === "delete" && options.role && (await deleteRole(iam, options));
+
+                    // * user assign commands (like in settings app)
+                    options.mode === "assign" &&
+                        options.user &&
+                        options.group &&
+                        (await assignUserToGroup(iam, options));
+                    options.mode === "assign" && options.user && options.role && (await assignUserToRole(iam, options));
+
+                    // * user remove commands (like in settings app)
+
+                    options.mode === "remove" &&
+                        options.user &&
+                        options.group &&
+                        (await removeUserFromGroup(iam, options));
+
+                    options.mode === "remove" &&
+                        options.role &&
+                        options.user &&
+                        (await removeRoleFromUser(iam, options));
+
+                    // * group assign commands (like in settings app)
+
+                    options.mode === "assign" &&
+                        options.group &&
+                        options.membergroup &&
+                        (await addMemberGroupToGroup(iam, options));
+
+                    options.mode === "assign" && options.group && options.role && (await addRoleToGroup(iam, options));
+
+                    // * group remove commands (like in settings app)
+
+                    options.mode === "remove" &&
+                        options.group &&
+                        options.membergroup &&
+                        (await removeMemberGroupFromGroup(iam, options));
+
+                    options.mode === "remove" &&
+                        options.group &&
+                        options.role &&
+                        (await removeRoleFromGroup(iam, options));
+
+                    // * role assign commands (like in settings app)
+
+                    options.mode === "assign" &&
+                        options.role &&
+                        options.memberrole &&
+                        (await addMemberRoleToRole(iam, options));
+
+                    // * role remove commands (like in settings app)
+
+                    options.mode === "remove" &&
+                        options.role &&
+                        options.memberrole &&
+                        (await removeMemberRoleFromRole(iam, options));
                 } catch (err) {
                     errorLog(err, options.verbose);
                 }
@@ -102,7 +161,7 @@ export default (program: Command) => {
             );
 
             console.log(
-                `    mc iam --mode info --user [user] --meta \t get all infos about ${color(
+                `\n    mc iam --mode info --user [user] --meta \t get all infos about ${color(
                     "users"
                 )} which contain ${color("[user]")} in the login name`
             );
@@ -119,16 +178,85 @@ export default (program: Command) => {
                 )} which contain ${roleColor("[role]")} in the name`
             );
 
-            // log(`\n    mc iam --mode create|delete --group groupName \t create or delete group`);
-            // log(`    mc iam --mode create|delete --user userName \t create or delete user`);
-            // log(`\n    mc iam --mode assign --group groupName --user userName \t\t Assign userName to groupName`);
-            // log(
-            //     `    mc iam --mode assign --group groupName --membergroup memberGroupName \t Assign memberGroupName to groupName`
-            // );
-            // log(`    mc iam --mode remove --group groupName --user userName \t\t Delete userName from groupName`);
-            // log(
-            //     `    mc iam --mode remove --group groupName --membergroup memberGroupName \t Delete memberGroupName from groupName`
-            // );
+            console.log(`\n    mc iam --mode create|delete --user <user> \t create or delete ${color("user")}`);
+            console.log(`    mc iam --mode create|delete --group <group>  create or delete ${groupColor("group")}`);
+            console.log(`    mc iam --mode create|delete --role  <role> \t create or delete ${roleColor("role")}`);
+            console.log(
+                `    mc iam --mode create --user <user> --subtenant subtenant \t create ${color(
+                    "user"
+                )} in specified subtenant`
+            );
+            console.log(
+                `\n    mc iam --mode assign --user <user> --group <group> \t assign ${color(
+                    "user"
+                )} to user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode assign --user <user> --role <role> \t assign ${roleColor("role")} to user ${color(
+                    "user"
+                )}`
+            );
+
+            console.log(
+                `    mc iam --mode assign --user <user> --group <group> \t assign ${color(
+                    "user"
+                )} to user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode assign --group <group> --membergroup <membergroup> \t assign ${groupColor(
+                    "membergroup"
+                )} to user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode assign --group <group> --role <role> \t\t\t assign ${roleColor(
+                    "role"
+                )} to user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode assign --role <role> --memberrole <memberrole> \t assign ${roleColor(
+                    "member role"
+                )} to custom role ${roleColor("role")}`
+            );
+
+            console.log(
+                `\n    mc iam --mode remove --user <user> --group <group> \t remove ${color(
+                    "user"
+                )} from user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode remove --user <user> --role <role> \t remove ${roleColor("role")} from user ${color(
+                    "user"
+                )}`
+            );
+
+            console.log(
+                `    mc iam --mode remove --user <user> --group <group> \t remove ${color(
+                    "user"
+                )} from user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode remove --group <group> --membergroup <membergroup> \t remove ${groupColor(
+                    "membergroup"
+                )} from user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode remove --group <group> --role <role> \t\t\t remove ${roleColor(
+                    "role"
+                )} from user group ${groupColor("group")}`
+            );
+
+            console.log(
+                `    mc iam --mode remove --role <role> --memberrole <memberrole> \t remove ${roleColor(
+                    "member role"
+                )} from custom role ${roleColor("role")}`
+            );
 
             serviceCredentialLog();
         });
@@ -171,6 +299,22 @@ function printGroups(
     console.log(`\n${groupColor(selected.length)} groups found`);
 }
 
+function printRoles(
+    users: IdentityManagementModels.ScimUserResponse[],
+    groups: IdentityManagementModels.ScimGroup[],
+    roles: IdentityManagementModels.ScimGroup[],
+    options: any
+) {
+    const selected =
+        typeof options.role === "string" ? roles.filter((role) => role.displayName?.includes(options.role)) : roles;
+
+    selected.forEach((role) => {
+        prettyPrintMembers(role, users, groups, roles, options);
+        options.meta && printObjectInfo("Meta:", role.meta || {}, options, ["lastModified"], color, 0);
+    });
+    console.log(`\n${roleColor(selected.length)} roles found.`);
+}
+
 function prettyPrintMembers(
     group: IdentityManagementModels.ScimGroup,
     users: IdentityManagementModels.ScimUserResponse[],
@@ -210,22 +354,6 @@ function prettyPrintMembers(
             console.log(`\t${roleColor(role)}`);
         });
     }
-}
-
-function printRoles(
-    users: IdentityManagementModels.ScimUserResponse[],
-    groups: IdentityManagementModels.ScimGroup[],
-    roles: IdentityManagementModels.ScimGroup[],
-    options: any
-) {
-    const selected =
-        typeof options.role === "string" ? roles.filter((role) => role.displayName?.includes(options.role)) : roles;
-
-    selected.forEach((role) => {
-        prettyPrintMembers(role, users, groups, roles, options);
-        options.meta && printObjectInfo("Meta:", role.meta || {}, options, ["lastModified"], color, 0);
-    });
-    console.log(`\n${roleColor(selected.length)} roles found.`);
 }
 
 async function getAllUsers(iam: IdentityManagementClient, options: any) {
@@ -344,114 +472,29 @@ function printUser(
         options.meta && printObjectInfo("Meta:", user || {}, options, ["lastModified"], color, 0);
     }
 }
-// async function listUsers(iam: IdentityManagementClient, options: any) {
-//     const filter: any = { attributes: "userName,groups,name" };
 
-//     if (options.user !== true) {
-//         filter.filter! = `userName sw "${options.user}"`;
-//     }
+async function createUser(iam: IdentityManagementClient, tm: TenantManagementClient, options: any) {
+    const newUser: any = { userName: options.user };
 
-//     const users = await iam.GetUsers(filter);
+    if (options.subtenant && typeof options.subtenant === "string") {
+        const subtenants = await tm.GetSubtenants();
+        const subtenant = subtenants.content?.find((x) => x.displayName === options.subtenant);
 
-//     let tenantAdmins = 0;
-//     users.resources?.forEach((user) => {
-//         const groups = user.groups?.filter((x) => x.display === "mdsp:core:TenantAdmin");
+        if (subtenant) {
+            newUser.subtenants = [{ id: subtenant.id }];
+        }
+    }
 
-//         const userColor = groups && groups.length ? color : (x: string) => x;
-//         const admin = groups && groups.length > 0 ? color("*") : "-";
-//         if (groups && groups.length > 0) tenantAdmins++;
-
-//         let userGroups: IdentityManagementModels.Group[] = [],
-//             userRoles: IdentityManagementModels.Group[] = [];
-
-//         user.groups?.forEach((grp) => {
-//             if (grp.display.startsWith("mdsp_usergroup:")) {
-//                 userGroups.push(grp);
-//             } else {
-//                 userRoles.push(grp);
-//             }
-//         });
-
-//         console.log(
-//             `${admin} ${userColor(user.userName)} [${groupColor(userGroups.length)} user groups, ${roleColor(
-//                 userRoles.length
-//             )} roles]`
-//         );
-
-//         if (options.verbose) {
-//             userGroups
-//                 .sort((x, y) => {
-//                     return x.display.localeCompare(y.display);
-//                 })
-//                 .forEach((x) => console.log(`\t - ${formatGroupOrRole(x.display)}`));
-//             userRoles
-//                 .sort((x, y) => {
-//                     return x.display.localeCompare(y.display);
-//                 })
-//                 .forEach((x) => console.log(`\t - ${formatGroupOrRole(x.display)}`));
-//         }
-//     });
-//     console.log(
-//         `Total Users: ${users.totalResults} total users. [${color(tenantAdmins + " tenant admins")}, ${
-//             (users.totalResults || 0) - tenantAdmins
-//         } other users.]`
-//     );
-// }
-
-// function formatGroupOrRole(groupName: string) {
-//     return groupName.startsWith("mdsp_usergroup:") ? groupColor(groupName) : roleColor(groupName);
-// }
-
-// async function listGroupsOrRoles(iam: IdentityManagementClient, options: any) {
-//     const filter: any = {};
-//     if (options.group) {
-//         filter.filter! = `displayName sw "${normalizeGroupName(options.group === true ? "" : options.group, options)}"`;
-//     }
-
-//     if (options.role && options.role !== true) {
-//         filter.filter! = `displayName sw "${options.role}"`;
-//     }
-
-//     let count = 0;
-//     const groups = await iam.GetGroups(filter);
-
-//     for await (const group of (groups.resources || []).sort((x, y) => {
-//         return x.displayName!.localeCompare(y.displayName!);
-//     })) {
-//         if (group.displayName!.startsWith("mdsp_usergroup:") && options.role) {
-//             continue;
-//         }
-//         count++;
-//         const userCount = options.verbose ? `[${group.members?.length} users]` : "";
-//         console.log(`${formatGroupOrRole(group.displayName!)} ${userCount}`);
-//         if (options.verbose) {
-//             for await (const member of group.members!) {
-//                 if (member.type === IdentityManagementModels.ScimGroupMember.TypeEnum.USER) {
-//                     const user = await iam.GetUser(member.value);
-//                     console.log(`\t ${user.userName}`);
-//                 } else {
-//                     const group = await iam.GetGroup(member.value);
-//                     console.log(`\t ${group.displayName}`);
-//                 }
-//             }
-//         }
-//     }
-
-//     if (options.group) {
-//         console.log(`Found: ${groupColor(count)} groups`);
-//     } else {
-//         console.log(`Found: ${roleColor(count)} roles`);
-//     }
-// }
-
-async function createUser(iam: IdentityManagementClient, options: any) {
-    const user = await iam.PostUser({ userName: options.user });
+    const user = await iam.PostUser(newUser);
     console.log(`user with username ${color(user.userName)} created`);
     verboseLog(JSON.stringify(user, null, 2), options.verbose);
 }
 
 async function createRole(iam: IdentityManagementClient, options: any) {
-    createGroup(iam, options);
+    const name = normalize(options.role, options);
+    const role = await iam.PostGroup({ displayName: name, description: `created using CLI` });
+    console.log(`custom role with displayName ${roleColor(role.displayName)} created`);
+    verboseLog(JSON.stringify(role, null, 2), options.verbose);
 }
 
 async function createGroup(iam: IdentityManagementClient, options: any) {
@@ -467,242 +510,409 @@ function normalize(name: string, options: any) {
             name = `mdsp_usergroup:${name}`;
         }
     } else if (options.role) {
-        if (!options.raw && !name.startsWith("mdsp_customrole:")) {
-            name = `mdsp_customrole:${name}`;
+        if (!options.raw && !name.startsWith("mdsp_customrole:") && !name.startsWith("mdsp:")) {
+            if (["Developer", "DeveloperAdmin", "StandardUser", "SubTenatnUser", "TenantAdmin"].includes(name)) {
+                name = "mdsp:core:" + name;
+            } else {
+                name = `mdsp_customrole:${name}`;
+            }
         }
     }
     return name;
 }
 
-function normalizeRoleName(name: string, options: any) {
-    if (name === "" || name === "true" || name === undefined) {
-        return "mdsp_usergroup:";
+async function deleteUser(iam: IdentityManagementClient, options: any) {
+    const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
+    if (users.totalResults === 1) {
+        const deletedUser = await iam.DeleteUser(users.resources![0].id!);
+        console.log(`user with username ${color(users.resources![0].userName!)} deleted`);
+        verboseLog(JSON.stringify(deletedUser, null, 2), options.verbose);
+    } else {
+        throwError(`found ${color(users.totalResults)} users users but expected 1 `);
     }
-    if (!options.raw && !name.startsWith("mdsp_userrole:")) {
-        name = `mdsp_usergroup:${name}`;
-    }
-    return name;
 }
 
-// async function deleteUser(iam: IdentityManagementClient, options: any) {
-//     const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
-//     if (users.totalResults === 1) {
-//         const deletedUser = await iam.DeleteUser(users.resources![0].id!);
-//         console.log(`user with username ${color(users.resources![0].userName!)} deleted`);
-//         verboseLog(JSON.stringify(deletedUser, null, 2), options.verbose);
-//     } else {
-//         throwError(`found ${color(users.totalResults)} users users but expected 1 `);
-//     }
-// }
+async function deleteGroup(iam: IdentityManagementClient, options: any) {
+    const groups = await iam.GetGroups({ filter: `displayName eq "${normalize(options.group, options)}"` });
+    if (groups.totalResults === 1) {
+        const deletedGroup = await iam.DeleteGroup(groups.resources![0].id!);
+        console.log(`group ${groupColor(groups.resources![0].displayName!)} deleted`);
+        verboseLog(JSON.stringify(deletedGroup, null, 2), options.verbose);
+    } else {
+        throwError(
+            `found ${color(
+                groups.totalResults
+            )} groups but expected 1 (--mode delete command expects the full name of the group in --group parameter) `
+        );
+    }
+}
 
-// async function deleteGroup(iam: IdentityManagementClient, options: any) {
-//     const groups = await iam.GetGroups({ filter: `displayName eq "${normalizeGroupName(options.group, options)}"` });
-//     if (groups.totalResults === 1) {
-//         const deletedGroup = await iam.DeleteGroup(groups.resources![0].id!);
-//         console.log(`group ${groupColor(groups.resources![0].displayName!)} deleted`);
-//         verboseLog(JSON.stringify(deletedGroup, null, 2), options.verbose);
-//     } else {
-//         throwError(`found ${color(groups.totalResults)} groups but expected 1 `);
-//     }
-// }
+async function deleteRole(iam: IdentityManagementClient, options: any) {
+    const roles = await iam.GetGroups({ filter: `displayName eq "${normalize(options.role, options)}"` });
+    if (roles.totalResults === 1) {
+        const deleteRole = await iam.DeleteGroup(roles.resources![0].id!);
+        console.log(`role ${roleColor(roles.resources![0].displayName!)} deleted`);
+        verboseLog(JSON.stringify(deleteRole, null, 2), options.verbose);
+    } else {
+        throwError(
+            `found ${color(
+                roles.totalResults
+            )} roles but expected 1 (--mode delete command expects the full name of the role in --role parameter) `
+        );
+    }
+}
 
-// async function assign(iam: IdentityManagementClient, options: any) {
-//     const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
-//     const groups = await iam.GetGroups({ filter: `displayName eq "${normalizeGroupName(options.group, options)}"` });
-//     const membergroups = await iam.GetGroups({
-//         filter: `displayName eq "${normalizeGroupName(options.membergroup, options)}"`,
-//     });
+async function assignUserToGroup(iam: IdentityManagementClient, options: any) {
+    const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
+    const groups = await iam.GetGroups({ filter: `displayName eq "${normalize(options.group, options)}"` });
 
-//     if (options.user && users.totalResults === 1 && groups.totalResults === 1) {
-//         const assigned = await iam.PostGroupMember(groups.resources![0].id!, {
-//             type: IdentityManagementModels.ScimGroupMember.TypeEnum.USER,
-//             value: users.resources![0].id!,
-//         });
-//         console.log(`assigned user ${color(options.user)} to ${color(normalizeGroupName(options.group, options))}`);
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.user) {
-//         throwError(
-//             `found ${color(users.totalResults)} users and ${color(groups.totalResults)} groups but expected 1 of each`
-//         );
-//     } else if (options.membergroup && membergroups.totalResults === 1 && groups.totalResults === 1) {
-//         const assigned = await iam.PostGroupMember(groups.resources![0].id!, {
-//             type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
-//             value: membergroups.resources![0].id!,
-//         });
-//         console.log(
-//             `assigned member ${color(options.membergroup)} to ${color(normalizeGroupName(options.group, options))}`
-//         );
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.membergroup) {
-//         throwError(
-//             `found ${color(membergroups.totalResults)} membergroups && ${color(
-//                 groups.totalResults
-//             )} groups but expected 1 of each`
-//         );
-//     }
-// }
+    if (users.totalResults !== 1) {
+        throwError(
+            `found ${color(users.totalResults)} users but expected 1 (you have to use the full --user name for ${
+                options.mode
+            } command)`
+        );
+    }
 
-// // the order of parameter is different for some reason in mindsphere when assigning groups and roles
-// // this is why we are using second method here to assign roles
-// // see also: https://github.com/mindsphere/mindconnect-nodejs/discussions/292
-// async function assignRole(iam: IdentityManagementClient, options: any) {
-//     const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
-//     const groups = await iam.GetGroups({ filter: `displayName eq "${normalizeGroupName(options.group, options)}"` });
-//     const roles = await iam.GetGroups({ filter: `displayName eq "${options.role}"` });
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
 
-//     // console.log(users.totalResults, groups.totalResults, roles.totalResults);
+    const assigned = await iam.PostGroupMember(groups.resources![0].id!, {
+        type: IdentityManagementModels.ScimGroupMember.TypeEnum.USER,
+        value: users.resources![0].id!,
+    });
+    console.log(`assigned user ${color(options.user)} to ${groupColor(normalize(options.group, options))}`);
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
 
-//     if (options.user && users.totalResults === 1 && roles.totalResults === 1) {
-//         const assigned = await iam.PostGroupMember(roles.resources![0].id!, {
-//             type: IdentityManagementModels.ScimGroupMember.TypeEnum.USER,
-//             value: users.resources![0].id!,
-//         });
-//         console.log(`Assigned role ${roleColor(options.role)} to ${color(options.user)} `);
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.user) {
-//         throwError(
-//             `found ${color(users.totalResults)} users and ${roleColor(roles.totalResults)} roles but expected 1 of each`
-//         );
-//     } else if (options.group && groups.totalResults === 1 && roles.totalResults === 1) {
-//         const assigned = await iam.PostGroupMember(roles.resources![0].id!, {
-//             type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
-//             value: groups.resources![0].id!,
-//         });
-//         console.log(`Assigned role ${roleColor(options.role)} to ${color(options.group)} `);
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.group) {
-//         throwError(
-//             `found ${color(groups.totalResults)} groups and ${roleColor(
-//                 roles.totalResults
-//             )} roles but expected 1 of each`
-//         );
-//     }
+async function assignUserToRole(iam: IdentityManagementClient, options: any) {
+    const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
+    const roles = await iam.GetGroups({ filter: `displayName eq "${normalize(options.role, options)}"` });
 
-//     //else if (options.role && roles.totalResults === 1) {
-//     //     const assigned = await iam.PostGroupMember(roles.resources![0].id!, {
-//     //         type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
-//     //         value: roles.resources![0].id!,
-//     //     });
-//     //     console.log(`assigned role ${color(options.role)} to ${color(normalizeGroupName(options.group, options))}`);
-//     //     verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     // } else if (options.role) {
-//     //     throwError(
-//     //         `found ${color(roles.totalResults)} roles and ${color(roles.totalResults)} roles but expected 1 of each`
-//     //     );
-//     // }
-// }
+    if (users.totalResults !== 1) {
+        throwError(
+            `found ${color(users.totalResults)} users but expected 1 (you have to use the full --user name for --mode ${
+                options.mode
+            } command)`
+        );
+    }
 
-// async function remove(iam: IdentityManagementClient, options: any) {
-//     const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
-//     const groups = await iam.GetGroups({ filter: `displayName eq "${normalizeGroupName(options.group, options)}"` });
-//     const membergroups = await iam.GetGroups({
-//         filter: `displayName eq "${normalizeGroupName(options.membergroup, options)}"`,
-//     });
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(
+                roleColor.totalResults
+            )} roles but expected 1 (you have to use the full --role name for --mode ${options.mode} command)`
+        );
+    }
 
-//     if (options.user && users.totalResults === 1 && groups.totalResults === 1) {
-//         const assigned = await iam.DeleteGroupMember(groups.resources![0].id!, users.resources![0].id!);
-//         console.log(`deleted user ${color(options.user)} from ${color(normalizeGroupName(options.group, options))}`);
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.user) {
-//         throwError(
-//             `found ${color(users.totalResults)} users and ${color(groups.totalResults)} groups but expected 1 of each`
-//         );
-//     } else if (options.membergroup && membergroups.totalResults === 1 && groups.totalResults === 1) {
-//         const assigned = await iam.DeleteGroupMember(groups.resources![0].id!, membergroups.resources![0].id!);
-//         console.log(
-//             `deleted member ${color(options.membergroup)} from ${color(normalizeGroupName(options.group, options))}`
-//         );
-//         verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
-//     } else if (options.membergroup) {
-//         throwError(
-//             `found ${color(membergroups.totalResults)} membergroups && ${color(
-//                 groups.totalResults
-//             )} groups but expected 1 of each`
-//         );
-//     }
-// }
+    const assigned = await iam.PostGroupMember(roles.resources![0].id!, {
+        type: IdentityManagementModels.ScimGroupMember.TypeEnum.USER,
+        value: users.resources![0].id!,
+    });
+    console.log(`assigned role ${roleColor(normalize(options.role, options))} to user ${color(options.user)}`);
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
 
-// function checkRequiredParameters(options: any) {
-//     !(["list", "create", "assign", "remove", "delete", "info"].indexOf(options.mode) >= 0) &&
-//         throwError(`invalid mode ${options.mode} (must be list, creete, assign, remove, delete, info)`);
+async function removeUserFromGroup(iam: IdentityManagementClient, options: any) {
+    const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
+    const groups = await iam.GetGroups({ filter: `displayName eq "${normalize(options.group, options)}"` });
 
-//     ["create", "delete"].forEach((x) => {
-//         options.mode === x &&
-//             options.role &&
-//             throwError(
-//                 `CREATE ROLE and DELETE ROLE operations are not public APIs in MindSphere. Please use the developer cockpit to create roles instead.`
-//             );
+    if (users.totalResults !== 1) {
+        throwError(
+            `found ${color(users.totalResults)} users but expected 1 (you have to use the full --user name for ${
+                options.mode
+            } command)`
+        );
+    }
 
-//         options.mode === x &&
-//             !options.user &&
-//             !options.group &&
-//             throwError(`you have to specify either --user [user] or --group [group] for mc iam --mode ${x} command`);
-//     });
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
 
-//     ["list", "info"].forEach((x) => {
-//         options.mode === x &&
-//             !options.user &&
-//             !options.group &&
-//             !options.role &&
-//             throwError(
-//                 `you have to specify either --user [user] or --group [group] or --role [role] for mc iam --mode ${x} command`
-//             );
+    const removed = await iam.DeleteGroupMember(groups.resources![0].id!, users.resources![0].id!);
+    console.log(`removed user ${color(options.user)} from ${groupColor(normalize(options.group, options))}`);
+    verboseLog(JSON.stringify(removed, null, 2), options.verbose);
+}
 
-//         options.mode === x &&
-//             +!!options.user + +!!options.group + +!!options.role > 1 &&
-//             throwError(
-//                 `you have to specify either --user [user] or --group [group] or --role [role]  for mc iam --mode ${x} command but not more than one`
-//             );
-//     });
+async function removeRoleFromUser(iam: IdentityManagementClient, options: any) {
+    const users = await iam.GetUsers({ filter: `userName eq "${options.user}"` });
+    const roles = await iam.GetGroups({ filter: `displayName eq "${normalize(options.role, options)}"` });
 
-//     ["info"].forEach((x) => {
-//         options.mode === x &&
-//             +!!options.user + +!!options.group + +!!options.role > 1 &&
-//             throwError(
-//                 `you have to specify either --user [user] or --group [group] or --role [role]  for mc iam --mode ${x} command but not more than one`
-//             );
-//     });
+    if (users.totalResults !== 1) {
+        throwError(
+            `found ${color(users.totalResults)} users but expected 1 (you have to use the full --user name for ${
+                options.mode
+            } command)`
+        );
+    }
 
-//     ["create", "delete", "assign", "remove"].forEach((x) => {
-//         options.mode === x &&
-//             options.user &&
-//             options.user === true &&
-//             throwError(`you have to specify full user name for iam --mode ${x} command`);
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(roles.totalResults)} groups but expected 1 (you have to use the full --group name for ${
+                options.mode
+            } command)`
+        );
+    }
 
-//         options.mode === x &&
-//             options.group &&
-//             options.group === true &&
-//             throwError(`you have to specify full group name for iam --mode ${x} command`);
+    const removed = await iam.DeleteGroupMember(roles.resources![0].id!, users.resources![0].id!);
+    console.log(`removed role ${roleColor(normalize(options.role, options))} from user ${color(options.user)}`);
+    verboseLog(JSON.stringify(removed, null, 2), options.verbose);
+}
 
-//         options.mode === x &&
-//             options.role &&
-//             options.role === true &&
-//             throwError(`you have to specify full role name for iam --mode ${x} command`);
-//     });
+async function addMemberGroupToGroup(iam: IdentityManagementClient, options: any) {
+    const groups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.group, { group: options.group })}"`,
+    });
+    const memberGroups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.membergroup, { group: options.membergroup })}"`,
+    });
 
-//     // ["assign", "remove"].forEach((x) => {
-//     //     options.mode === x &&
-//     //         (options.group === true || !options.group) &&
-//     //         throwError(`you have to specify --group [group] iam --mode ${x} command`);
+    if (memberGroups.totalResults !== 1) {
+        throwError(
+            `found ${color(
+                memberGroups.totalResults
+            )} member groups but expected 1 (you have to use the full --membergroup name for ${options.mode} command)`
+        );
+    }
 
-//     //     options.mode === x &&
-//     //         options.user &&
-//     //         options.membergroup &&
-//     //         throwError(
-//     //             `you have to specify --user [user] or --membergroup [membergroup] iam --mode ${x} command but not both`
-//     //         );
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
 
-//     //     options.mode === x &&
-//     //         !options.user &&
-//     //         !options.membergroup &&
-//     //         throwError(
-//     //             `you have to specify either --user [user] or --membergroup [membergroup] iam --mode ${x} command `
-//     //         );
+    const assigned = await iam.PostGroupMember(groups.resources![0].id!, {
+        type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
+        value: memberGroups.resources![0].id!,
+    });
+    console.log(
+        `assigned memberGroup ${groupColor(
+            normalize(options.membergroup, { group: options.membergroup })
+        )} to ${groupColor(normalize(options.group, { group: options.group }))}`
+    );
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
 
-//     //     options.mode === x &&
-//     //         (options.user === true || options.membergroup === true) &&
-//     //         throwError(
-//     //             `you have to specify either --user [user] or --membergroup [membergroup] iam --mode ${x} command  (no empty parameters)`
-//     //         );
-//     // });
-// }
+async function removeMemberGroupFromGroup(iam: IdentityManagementClient, options: any) {
+    const groups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.group, { group: options.group })}"`,
+    });
+    const memberGroups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.membergroup, { group: options.membergroup })}"`,
+    });
+
+    if (memberGroups.totalResults !== 1) {
+        throwError(
+            `found ${color(
+                memberGroups.totalResults
+            )} member groups but expected 1 (you have to use the full --membergroup name for ${options.mode} command)`
+        );
+    }
+
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
+
+    const assigned = await iam.DeleteGroupMember(groups.resources![0].id!, memberGroups.resources![0].id!);
+    console.log(
+        `removed memberGroup ${groupColor(
+            normalize(options.membergroup, { group: options.membergroup })
+        )} from ${groupColor(normalize(options.group, { group: options.group }))}`
+    );
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
+
+async function addRoleToGroup(iam: IdentityManagementClient, options: any) {
+    const groups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.group, { group: options.group })}"`,
+    });
+    const roles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.role, { role: options.role })}"`,
+    });
+
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(roles.totalResults)} roles but expected 1 (you have to use the full --role name for ${
+                options.mode
+            } command)`
+        );
+    }
+
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
+
+    const assigned = await iam.PostGroupMember(roles.resources![0].id!, {
+        type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
+        value: groups.resources![0].id!,
+    });
+    console.log(
+        `assigned role ${roleColor(normalize(options.role, { role: options.role }))} to group ${groupColor(
+            normalize(options.group, { group: options.group })
+        )}`
+    );
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
+
+async function removeRoleFromGroup(iam: IdentityManagementClient, options: any) {
+    const groups = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.group, { group: options.group })}"`,
+    });
+    const roles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.role, { role: options.role })}"`,
+    });
+
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(roles.totalResults)} roles but expected 1 (you have to use the full --role name for ${
+                options.mode
+            } command)`
+        );
+    }
+
+    if (groups.totalResults !== 1) {
+        throwError(
+            `found ${groupColor(
+                groups.totalResults
+            )} groups but expected 1 (you have to use the full --group name for ${options.mode} command)`
+        );
+    }
+
+    const assigned = await iam.DeleteGroupMember(roles.resources![0].id!, groups.resources![0].id!);
+    console.log(
+        `removed role ${roleColor(normalize(options.role, { role: options.role }))} from group ${groupColor(
+            normalize(options.group, { group: options.group })
+        )}`
+    );
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
+
+async function addMemberRoleToRole(iam: IdentityManagementClient, options: any) {
+    const roles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.role, { role: options.role })}"`,
+    });
+    const memberRoles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.memberrole, { role: options.role })}"`,
+    });
+
+    if (memberRoles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(
+                memberRoles.totalResults
+            )} member roles but expected 1 (you have to use the full --memberrole name for ${options.mode} command)`
+        );
+    }
+
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(roles.totalResults)} roles but expected 1 (you have to use the full --role name for ${
+                options.mode
+            } command)`
+        );
+    }
+
+    const assigned = await iam.PostGroupMember(memberRoles.resources![0].id!, {
+        type: IdentityManagementModels.ScimGroupMember.TypeEnum.GROUP,
+        value: roles.resources![0].id!,
+    });
+    console.log(
+        `assigned member role ${roleColor(normalize(options.memberrole, { group: options.memberrole }))} to ${roleColor(
+            normalize(options.role, { group: options.role })
+        )}`
+    );
+    verboseLog(JSON.stringify(assigned, null, 2), options.verbose);
+}
+
+async function removeMemberRoleFromRole(iam: IdentityManagementClient, options: any) {
+    const roles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.role, { role: options.role })}"`,
+    });
+    const memberRoles = await iam.GetGroups({
+        filter: `displayName eq "${normalize(options.memberrole, { role: options.role })}"`,
+    });
+
+    if (memberRoles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(
+                memberRoles.totalResults
+            )} member roles but expected 1 (you have to use the full --memberrole name for ${options.mode} command)`
+        );
+    }
+
+    if (roles.totalResults !== 1) {
+        throwError(
+            `found ${roleColor(roles.totalResults)} roles but expected 1 (you have to use the full --role name for ${
+                options.mode
+            } command)`
+        );
+    }
+
+    const removed = await iam.DeleteGroupMember(memberRoles.resources![0].id!, roles.resources![0].id!);
+    console.log(
+        `removed member role ${roleColor(
+            normalize(options.memberrole, { group: options.memberrole })
+        )} from ${roleColor(normalize(options.role, { group: options.role }))}`
+    );
+    verboseLog(JSON.stringify(removed, null, 2), options.verbose);
+}
+
+function checkRequiredParameters(options: any) {
+    !(["list", "create", "assign", "remove", "delete", "info"].indexOf(options.mode) >= 0) &&
+        throwError(`invalid mode ${options.mode} (must be list, creete, assign, remove, delete, info)`);
+
+    ["list", "info"].forEach((x) => {
+        options.mode === x &&
+            !options.user &&
+            !options.group &&
+            !options.role &&
+            throwError(
+                `you have to specify either --user [user] or --group [group] or --role [role] for mc iam --mode ${x} command`
+            );
+    });
+
+    ["create", "delete"].forEach((x) => {
+        options.mode === x &&
+            !(typeof options.user === "string") &&
+            !(typeof options.group === "string") &&
+            !(typeof options.role === "string") &&
+            throwError(
+                `you have to specify either --user [user] or --group [group] or --role [role] for mc iam --mode ${x} command`
+            );
+    });
+
+    ["create", "delete", "assign", "remove"].forEach((x) => {
+        options.mode === x &&
+            options.user &&
+            options.user === true &&
+            throwError(`you have to specify full user name for iam --mode ${x} command`);
+
+        options.mode === x &&
+            options.group &&
+            options.group === true &&
+            throwError(`you have to specify full group name for iam --mode ${x} command`);
+
+        options.mode === x &&
+            options.role &&
+            options.role === true &&
+            throwError(`you have to specify full role name for iam --mode ${x} command`);
+    });
+}